+++ libsignal-service/src/groups_v2/mod.rs
@@ -1,11 +1,9 @@
 //! Everything needed to support [Signal Groups v2](https://signal.org/blog/new-groups/)
-mod api;
-mod models;
+mod manager;
 mod operations;
 
-pub use api::{
-    CredentialsCache, CredentialsCacheError, GroupsV2Api,
+pub use manager::{
+    CredentialsCache, CredentialsCacheError, GroupsManager,
     InMemoryCredentialsCache,
 };
 pub use operations::GroupDecryptionError;
-pub use zkgroup::auth::AuthCredentialResponse;
